Added main file
Pretty much the only code for this
- uses Arduino Analog read for ADC functions
- uses Arduino Serial functions for exporting over UART

The soon to be implemented command shell will run in the main loop in order to receive commands at a lower priority than the periodic measurements. 

An interrupt is used to log measurements at a periodic rate. The interrupt will trigger for 30 ticks, and start logging for the last 10 ticks.
 The ADC will take 10 measurements and average the measurements and report that result.

TODO: 

Implement moving average to filter out errors in measurement

Add time stamp to log (seconds)

Implement command shell so datalog can be left and downloaded later