Updated Interrupt
Updated the interrupt sequence so all ADC processing is done outside the interrupt to avoid (unlikely) concurrency issues. 
While this was not necessary it was done for good practice.

changed all timing to be based off of the main counter as to limit global variables, this does come at the price of performing more operations. Not sure which is more ideal but for this scenario it is irrelevant.